requirements.txt
numpy
scipy
scikit-learn
umap-learn
matplotlib
plotly
bokeh
uproot
awkward
mplhep